Added functionality to printer class and added a test case in main (UNTESTED BUT COMPILES)
added code so printer class can print a string with the startPrint
method. Don't know if this works because I don't have a printer to test.
It should format the string to fit on one line but it's all in theory.
Also the startPrint method formats the string to be the paper sized
specified by the client.