Changed GUI a little and other small changes
Added ability for the user to specify port number. This is optional
because the user might not need this option. Changed Printer constructer
to be a method incase the user doesn't select a printer. If the user
doesn't select a printer then the program will tell the user and won't
start until a printer is chosen.